third_party/WebKit/LayoutTests/animations/display-change-does-not-terminate-animation-expected.txt

Tests that changing the display property of a running animation to a value other than 'none' does not cause it to terminate or re-start.

target
INFO: Start event fired
INFO: Setting display to inline
PASS: Left property was not reset
INFO: Setting display to block
PASS: Left property was not reset
INFO: Setting display to inline-block
PASS: Left property was not reset
INFO: Setting display to inline-table
PASS: Left property was not reset
INFO: Setting display to list-item
PASS: Left property was not reset
INFO: Setting display to run-in
PASS: Left property was not reset
INFO: Setting display to table
PASS: Left property was not reset
INFO: Setting display to table-caption
PASS: Left property was not reset
INFO: Setting display to table-column-group
PASS: Left property was not reset
INFO: Setting display to table-header-group
PASS: Left property was not reset
INFO: Setting display to table-footer-group
PASS: Left property was not reset
INFO: Setting display to table-row-group
PASS: Left property was not reset
INFO: Setting display to table-cell
PASS: Left property was not reset
INFO: Setting display to table-column
PASS: Left property was not reset
INFO: Setting display to table-row
PASS: Left property was not reset
